+++ CHANGES.rst
@@ -12,6 +12,9 @@ Changelog
     The plone.app.event.dx.event type and plone.app.event.dx:default profile
     are deprecated and will be removed in a future version of plone.app.event.
 
+- Fix icalendar export for collections and Archetype topics. Fixes #104.
+  [thet]
+
 - Don't include occurrences in icalendar exports of event_listing, but include
   the original event with it's recurrence rule. Fixes #103.
   [thet]